created Game_Controller class and broke up Game.start_game method
